refactor (reddit_search): refactor the folder structure
