+++ mcp_servers/reddit_search/tools/__init__.py
@@ -1,0 +1,17 @@
+# Module exports and imports
+from .find_subreddits import find_relevant_subreddits
+from .search_posts import search_subreddit_posts
+from .get_comments import get_post_and_top_comments
+from .find_similar import find_similar_posts_reddit
+from .shared_utils import SubredditInfo, PostInfo, CommentInfo, PostDetails
+
+__all__ = [
+    "find_relevant_subreddits",
+    "search_subreddit_posts", 
+    "get_post_and_top_comments",
+    "find_similar_posts_reddit",
+    "SubredditInfo",
+    "PostInfo", 
+    "CommentInfo",
+    "PostDetails"
+]
